Tank assets increased (thinking about making this the main focus) , refined other assets and the kitchen now a room instead of a table. Trying to make sure I actually set up my objects properly things time! haha
Got carrier away with shiny things in general but happy I did none the less. My hands are faster navigating and playing with the system. there is also a bunch of assets that included water movement - I did use that, but will make sure to cite anyone and everyone used :)